Add files via upload
Articly.py is the article class I used. 

HashTable.py is the hash table that backs up the map. It has a load factor of 0.7, a resizing factor of 2, and initial size of 10. It hashes based on keyword. It uses quadratic probing to resolve collisions. Buckets are lists of Article objects. Each list contains article objects that all have the same keyword. Attempts to access a certain keyword will return all the Article objects with that keyword or an empty list if that keyword isn't in the hash table.

unordered_map.py is a map that will be used to store everything. The actual map part is an instance of the HashTable class. The key passed into the map to insert is irrelevant, but the value MUST be an Article object. The size of the map is NOT the size of the hash table, it is the number of article objects in the hash table. It supports the use of brackets to insert and access such as ("myMap[keyword] = ArticleObj" or "resultList = myMap[keyword]"). It also supports the use of "in" to check if a keyword is in the map. ex.) "'politics' in myMap" returns true if "politics" is a keyword of some article in the map or returns false if it isn't. ClearMap() clears the map.