Add files via upload
Added function InsertAll(articleList) that takes a list of Article objects as a parameter and inserts them all into the unordered map. 
Optimized the insertion function in the hashtable so that when we increase table size we only rehash the sub-lists, not every article in the sub-list. This works because we just move the entire sub-list to the correct spot in the table instead of individually moving every article to the correct spot in the table.